requirements.txt
torch~=1.8.1+cu111
matplotlib~=3.3.1
pandas~=1.2.3
numpy~=1.19.1
Pillow~=7.2.0
torchvision~=0.9.1+cu111
opencv-python~=4.5.2.52